Break up ScrollLockProvider
